+++ Cargo.toml
@@ -12,16 +12,17 @@ repository = ""
 [workspace.dependencies]
 
 # kona
-kona-common = { git = "https://github.com/moongate-forks/kona/", branch = "kona-upstream" }
-kona-common-proc = { git = "https://github.com/moongate-forks/kona/", branch = "kona-upstream" }
-kona-preimage = { git = "https://github.com/moongate-forks/kona/", branch = "kona-upstream", features = [
+kona-common = { git = "https://github.com/moongate-forks/kona/", branch = "new-upstream" }
+kona-common-proc = { git = "https://github.com/moongate-forks/kona/", branch = "new-upstream" }
+kona-preimage = { git = "https://github.com/moongate-forks/kona/", branch = "new-upstream", features = [
     "rkyv",
 ] }
-kona-primitives = { git = "https://github.com/moongate-forks/kona/", branch = "kona-upstream" }
-kona-mpt = { git = "https://github.com/moongate-forks/kona/", branch = "kona-upstream" }
-kona-derive = { git = "https://github.com/moongate-forks/kona/", branch = "kona-upstream", default-features = false }
-kona-executor = { git = "https://github.com/moongate-forks/kona/", branch = "kona-upstream" }
-kona-client = { git = "https://github.com/moongate-forks/kona/", branch = "kona-upstream" }
+kona-primitives = { git = "https://github.com/moongate-forks/kona/", branch = "new-upstream" }
+kona-mpt = { git = "https://github.com/moongate-forks/kona/", branch = "new-upstream" }
+kona-derive = { git = "https://github.com/moongate-forks/kona/", branch = "new-upstream", default-features = false }
+kona-executor = { git = "https://github.com/moongate-forks/kona/", branch = "new-upstream" }
+kona-client = { git = "https://github.com/moongate-forks/kona/", branch = "new-upstream" }
+kona-host = { git = "https://github.com/moongate-forks/kona/", branch = "new-upstream" }
 
 # general
 anyhow = { version = "1.0.86", default-features = false }
@@ -56,7 +57,6 @@ alloy-primitives = { version = "0.7.6", default-features = false }
 alloy-rlp = { version = "0.3.5", default-features = false }
 alloy-consensus = { version = "0.1", default-features = false }
 alloy-sol-types = { version = "0.7.6" }
-kona-host = { git = "https://github.com/moongate-forks/kona/", branch = "kona-upstream" }
 op-alloy-consensus = { version = "0.1.0", default-features = false }
 alloy-eips = { version = "0.1", default-features = false }
 revm = { git = "https://github.com/bluealloy/revm", tag = "v37", version = "10.0.0", default-features = false }
@@ -69,11 +69,6 @@ sp1-helper = { git = "https://github.com/succinctlabs/sp1.git", branch = "dev" }
 
 [profile.release-client-lto]
 inherits = "release"
-panic = "abort"
-codegen-units = 1
-lto = "fat"
-
-[profile.release]
 panic = "abort"
 codegen-units = 1
 lto = "fat"
